重构检测结果转换逻辑
在 `imet_ModelEngine_yolov11_obb.hpp` 中，删除了 `Detection` 类的转换运算符，并添加了 `convertDetectionToDetectionRectangleInfo` 方法以处理检测结果的转换。在 `imet_ModelEngine_yolov11_obb.cpp` 中，移除了该运算符的实现，改为使用新的转换方法，并进行了缩放处理。

在 `ime_utilty.hpp` 中，修改了 `Point` 类型为整数对，并将 `PainterConfig` 中的 `shapeType` 默认值改为 `Circle`。同时，删除了一些与绘制相关的静态方法，以简化接口和重构绘制逻辑。

+++ ime/utilty/include/ime_utilty.hpp
@@ -5,7 +5,7 @@
 #include"opencv2/opencv.hpp"
 
 namespace rw {
-	using Point = std::pair<double, double>;
+	using Point = std::pair<int, int>;
 
     struct DetectionRectangleInfo
     {
@@ -36,7 +36,7 @@ namespace rw {
 
         struct PainterConfig
         {
-            ShapeType shapeType{ ShapeType ::Rectangle};
+            ShapeType shapeType{ ShapeType::Rectangle};
             int fontSize=5;
             int fontThickness = 1;
             int thickness=2;
@@ -57,14 +57,11 @@ namespace rw {
             PainterConfig config
         );
 
-
         static cv::Mat drawShapes(
             const cv::Mat& image,
             const DetectionRectangleInfo& rectInfo,
             PainterConfig config
         );
-
-
 
         static void drawShapesOnSourceImg(
             cv::Mat& image,
